% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{AIC.GAMBiCop}
\alias{AIC,GAMBiCop-method}
\alias{AIC.GAMBiCop}
\title{Akaike's 'An Information Criterion' for a fitted \code{\link{GAMBiCop-class}}}
\usage{
AIC.GAMBiCop(object, ..., k = 2)

\S4method{AIC}{GAMBiCop}(object, ..., k = 2)
}
\arguments{
\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}

\item{k}{numeric, the penalty per parameter to be used; the default \code{k = 2} is the classical AIC.}
}
\value{
A numeric value with the corresponding AIC.
}
\description{
Function calculating Akaike's 'An Information Criterion' (AIC) for a fitted \code{\link{GAMBiCop-class}}
object (note that the models are usually fitted by penalized likelihood maximization).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

