% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{AIC.gamBiCop}
\alias{AIC.gamBiCop}
\alias{AIC,gamBiCop-method}
\title{Akaike's An Information Criterion for a gamBiCop Object}
\usage{
\S4method{AIC}{gamBiCop}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{un-used in this class}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}
}
\value{
A numeric value with the corresponding AIC.
}
\description{
Function calculating Akaike's 'An Information Criterion' (AIC) for an object
of the class \code{\link[gamCopula:gamBiCop-class]{gamBiCop}}
(note that the models are usually fitted by penalized likelihood
maximization).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}
