% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{BIC.GAMBiCop}
\alias{BIC,GAMBiCop-method}
\alias{BIC.GAMBiCop}
\title{Schwarz's Bayesian Information Criterion for a fitted \code{\link{GAMBiCop-class}}}
\usage{
BIC.GAMBiCop(object, ...)

\S4method{BIC}{GAMBiCop}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
A numeric value with the corresponding BIC.
}
\description{
Function calculating the Schwarz's Bayesian Information Criterion (BIC)
for a fitted \code{\link{GAMBiCop-class}} object
(note that the models are usually fitted by penalized likelihood maximization).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

