% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{BIC.gamBiCop}
\alias{BIC.gamBiCop}
\alias{BIC,gamBiCop-method}
\title{Schwarz's Bayesian Information Criterion for a gamBiCop Object}
\usage{
\S4method{BIC}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{un-used in this class}
}
\value{
A numeric value with the corresponding BIC.
}
\description{
Function calculating the Schwarz's Bayesian Information Criterion (BIC)
for an object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}} (note that the models are
usually fitted by penalized likelihood maximization).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}
