% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilsPublic.R
\name{BiCopEta2Par}
\alias{BiCopEta2Par}
\title{Copula Parameter of a Bivariate Copula for a Given Value of the Calibration
Function}
\usage{
BiCopEta2Par(family, eta)
}
\arguments{
\item{family}{A copula family:
\code{1} Gaussian,
\code{2} Student t,
\code{3} Clayton,
\code{4} Gumbel,
\code{13} Survival Clayton,
\code{14} Survival Gumbel,
\code{23} Rotated (90 degrees) Clayton,
\code{24} Rotated (90 degrees) Gumbel,
\code{33} Rotated (270 degrees) Clayton and
 \code{34} Rotated (270 degrees) Gumbel.}

\item{par}{The (first) copula parameter}
}
\value{
The value of the calibration function, depending on the copula
parameter and family as:
\itemize{
\item \code{1} Gaussian, \code{f(x) = tanh(x/2)}
\item \code{2} Student t, \code{f(x) = tanh(x/2)}
\item \code{3} Clayton, \code{f(x) = exp(x)}
\item \code{4} Gumbel, \code{f(x) = 1+exp(x)}
\item \code{13} Survival Clayton, \code{f(x) = exp(x)}
\item \code{14} Survival Gumbel, \code{f(x) = 1+exp(x)}
\item \code{23} Rotated (90 degrees) Clayton, \code{f(x) = -exp(x)}
\item \code{24} Rotated (90 degrees) Gumbel, \code{f(x) = -1-exp(x)}
\item \code{33} Rotated (270 degrees) Clayton, \code{f(x) = -exp(x)}
\item \code{34} Rotated (270 degrees) Gumbel, \code{f(x) = -1-exp(x)}
}
}
\description{
Computes the (first) copula parameter of a bivariate copula for a given
value of the calibration function (eta).
}
\seealso{
\code{\link{BiCopEta2Par}} or \code{\link{BiCopPar2Tau}} and
\code{\link{BiCopTau2Par}} from
\code{\link[VineCopula:VineCopula-package]{VineCopula}}.
}

