% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilsPublic.R
\name{BiCopPar2Eta}
\alias{BiCopPar2Eta}
\title{Calibration Function of a Bivariate Copula for a Given Parameter's Value}
\usage{
BiCopPar2Eta(family, par)
}
\arguments{
\item{family}{A copula family:
\code{1} Gaussian,
\code{2} Student t,
\code{3} Clayton,
\code{4} Gumbel,
\code{13} Survival Clayton,
\code{14} Survival Gumbel,
\code{23} Rotated (90 degrees) Clayton,
\code{24} Rotated (90 degrees) Gumbel,
\code{33} Rotated (270 degrees) Clayton and
 \code{34} Rotated (270 degrees) Gumbel.}

\item{par}{The (first) copula parameter}
}
\value{
The value of the calibration function, depending on the copula
parameter and family as:
\itemize{
\item \code{1} Gaussian, \code{f(x) = 2*atanh(x)}
\item \code{2} Student t, \code{f(x) = 2*atanh(x)}
\item \code{3} Clayton, \code{f(x) = log(x)}
\item \code{4} Gumbel, \code{f(x) = log(x-1)}
\item \code{13} Survival Clayton, \code{f(x) = log(x)}
\item \code{14} Survival Gumbel, \code{f(x) = log(x-1)}
\item \code{23} Rotated (90 degrees) Clayton, \code{f(x) = -log(x)}
\item \code{24} Rotated (90 degrees) Gumbel, \code{f(x) = -log(x-1)}
\item \code{33} Rotated (270 degrees) Clayton, \code{f(x) = -log(x)}
\item \code{34} Rotated (270 degrees) Gumbel, \code{f(x) = -log(x-1)}
}
}
\description{
Computes the calibration function (eta) of a bivariate copula for a given
value of the (first) copula parameter.
}
\seealso{
\code{\link{BiCopEta2Par}} or \code{\link{BiCopPar2Tau}} and
\code{\link{BiCopTau2Par}} from
\code{\link[VineCopula:VineCopula-package]{VineCopula}}.
}

