% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{dim.gamVineCop}
\alias{dim,gamVineCop-method}
\alias{dim.gamVineCop}
\title{Dimension of a \code{gamVineCop-class} object}
\usage{
\method{dim}{gamVineCop}(x)

\S4method{dim}{gamVineCop}(x)
}
\arguments{
\item{x}{fitted \code{\link{gamVineCop-class}} object.}
}
\value{
Dimension of the \code{gamVineCop-class} object.
}
\description{
Retrieve the dimension of a \code{gamVineCop-class} object.
}
\seealso{
\code{gamVineCop-class} and \code{gamVineCop}.
}

