% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\docType{methods}
\name{dim.gamVine}
\alias{dim,gamVine-method}
\alias{dim.gamVine}
\title{Dimension of an Object of the Class gamVine}
\usage{
\S4method{dim}{gamVine}(x)
}
\arguments{
\item{x}{An object of the class 
\code{\link[gamCopula:gamVine-class]{gamVine}}.}
}
\value{
Dimension of the 
\code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Retrieve the dimension of an object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}

