% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{EDF.gamBiCop}
\alias{EDF.gamBiCop}
\title{Equivalent Degrees of Freedom for a fitted \code{\link{gamBiCop-class}}}
\usage{
EDF.gamBiCop(object)
}
\arguments{
\item{object}{fitted \code{\link{gamBiCop-class}} object.}
}
\value{
Estimated degrees of freedom for each smooth component.
}
\description{
Function calculating the Equivalent Degrees of Freedom (EDF)
for a fitted \code{\link{gamBiCop-class}} object.
It basically sums the edf of the \code{\link{gamObject}}
for each smooth component.
}

