% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{formula.GAMBiCop}
\alias{formula,GAMBiCop-method}
\alias{formula.GAMBiCop}
\title{\code{\link{GAMBiCop-class}} formula}
\usage{
formula.GAMBiCop(x, ...)

\S4method{formula}{GAMBiCop}(x, ...)

formula.GAMBiCop(x, ...)

\S4method{formula}{GAMBiCop}(x, ...)
}
\arguments{
\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}

\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}

\item{x}{\R object.}
}
\description{
Description of the \code{\link{gam}} formula for  fitted \code{\link{GAMBiCop-class}} object.
This function is a wrapper to \code{\link{formula.gam}}
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.

Description of the \code{\link{gam}} formula for  fitted \code{\link{GAMBiCop-class}} object.
This function is a wrapper to \code{\link{formula.gam}}
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\seealso{
\code{\link{formula.gam}} function
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.

\code{\link{formula.gam}} function
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}

