% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamBiCopEst}
\alias{gamBiCopEst}
\title{Estimate a Generalized Additive model (gam) for the copula parameter or Kendall's tau}
\usage{
gamBiCopEst(family = 1, parFrhs = ~1, dataset, tau = FALSE,
  method = "FS", tol.rel = 0.001, n.iters = 10, verbose = FALSE)
}
\arguments{
\item{family}{A copula family: \code{1} Gaussian, \code{2} Student t,
\code{3} Clayton, \code{4} Gumbel, \code{13} Survival Clayton, \code{14} Survival Gumbel,
\code{23} Rotated (90 degrees) Clayton, \code{24} Rotated (90 degrees) Gumbel,
\code{33} Rotated (270 degrees) Clayton and \code{34} Rotated (270 degrees) Gumbel.}

\item{parFrhs}{A gam formula (see \code{\link{gam}}, \code{\link{formula.gam}} and \code{\link{gam.models}} from the \code{\link[mgcv:mgcv-package]{mgcv}} package).}

\item{dataset}{A matrix or data frame containing the model responses (in [0,1]x[0,1]) and covariates required by the formula.}

\item{tau}{\code{FALSE} (default) for a calibration fonction specified for the Copula parameter
or \code{TRUE} for a calibration function specified for Kendall's tau.}

\item{method}{\code{"FS"} for Fisher-scoring and \code{"NR"} for Newton-Raphson.}

\item{tol.rel}{Relative tolerance for \code{"FS"}/\code{"NR"} algorithm.}

\item{n.iters}{Maximal number of iterations for \code{"FS"}/\code{"NR"} algorithm.}

\item{verbose}{\code{TRUE} prints informations during the estimation.}
}
\value{
\code{gamBiCopEst} returns a list consisting of
\item{res}{S4 \code{\link{gamBiCop-class}} object.}
\item{method}{\code{"FS"} for Fisher-scoring and \code{"NR"} for Newton-Raphson.}
\item{tol.rel}{relative tolerance for \code{"FS"}/\code{"NR"} algorithm.}
\item{n.iters}{maximal number of iterations for \code{"FS"}/\code{"NR"} algorithm.}
\item{trace}{the estimation procedure's trace.}
}
\description{
Estimate a Generalized Additive model (gam) for the copula parameter or Kendall's tau
}
\seealso{
\code{\link{gamBiCop}} and \code{\link{gamBiCopEst}}.
}

