% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamBiCopPred}
\alias{gamBiCopPred}
\title{Predict from fitted \code{\link{gamBiCop-class}} object}
\usage{
gamBiCopPred(object, newdata = NULL, target = "calib", alpha = 0,
  type = "link")
}
\arguments{
\item{object}{fitted \code{\link{gamBiCop-class}} object.}

\item{newdata}{(Same as in \code{\link{predict.gam}} from the \code{\link[mgcv:mgcv-package]{mgcv}} package) A matrix or data frame containing the values of the model covariates at which predictions are required.
If this is not provided then predictions corresponding to the original data are returned.
If newdata is provided then it should contain all the variables needed for prediction:
a warning is generated if not.}

\item{target}{Either \code{"calib"}, \code{"par"} or \code{"tau"} or a combination of those.
\code{"calib"} (default) corresponds to the calibration function,
\code{"par"} to the copula parameter and \code{"tau"} to Kendall's tau.}

\item{alpha}{In (0,1) to return the corresponding confidence interval.}

\item{type}{(Similar as in \code{\link{predict.gam}} from the \code{\link[mgcv:mgcv-package]{mgcv}} package,
only active for \code{type = "calib"}). When this has the value \code{"link"} (default), the calibration function is returned.
 When \code{type = "terms"} each component of the linear predictor is returned seperately (possibly with standard errors):
    this includes parametric model components, followed by each smooth component, but excludes any offset and any intercept.
    When \code{type = "lpmatrix"} then a matrix is returned which yields the values of the linear predictor (minus any offset)
    when postmultiplied by the parameter vector (in this case alpha is ignored).}
}
\value{
If \code{target = "calib"}, then a list with 1 item \code{calib}.
If \code{target = "par"}, \code{target = "tau"} or \code{target = c("par", "tau")},
then a list with 2, 2 or 3 items, namely \code{calib} and \code{par},  \code{tau} and \code{par},
or  \code{calib}, \code{tau} and \code{par}.

If \code{alpha} is in (0,1), then a additional items of the list are \code{calib.CI}
 as well as e.g. \code{par.CI} and/or \code{tau.CI} depending on the value of \code{target}.

Otherwhise, if \code{type = "lpmatrix"} (only active for \code{type = "calib"}),
then a matrix is returned which will give a vector of linear predictor values (minus any offest) at the supplied covariate
values, when applied to the model coefficient vector (similar as \code{\link{predict.gam}} from the \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\description{
Predict from fitted \code{\link{gamBiCop-class}} object
}
\seealso{
\code{\link{gamBiCop}} and \code{\link{gamBiCopPred}}.
}

