% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamBiCopSim}
\alias{gamBiCopSim}
\title{Simulate from fitted \code{\link{gamBiCop-class}} object}
\usage{
gamBiCopSim(object, newdata = NULL, N = NULL, return.calib = FALSE,
  return.par = FALSE, return.tau = FALSE)
}
\arguments{
\item{object}{fitted \code{\link{gamBiCop-class}} object.}

\item{newdata}{(same as in \code{\link{predict.gam}} from the \code{\link[mgcv:mgcv-package]{mgcv}} package) A matrix or data frame containing the values of the model covariates at which simulations are required.
If this is not provided then simulations corresponding to the original data are returned.}

\item{N}{sample size.}

\item{return.calib}{should the calibration function (\code{TRUE}) be returned or not (\code{FALSE})?}

\item{return.par}{should the copula parameter (\code{TRUE}) be returned or not (\code{FALSE})?}

\item{return.tau}{should the Kendall's tau (\code{TRUE}) be returned or not (\code{FALSE})?}
}
\value{
A list with 1 item \code{data}.  When \code{N} is smaller or larger than the \code{newdata}'s number of rows
(or the number of rows in the original data if \code{newdata} is not provided),
then \code{N} observations are sampled uniformly (with replacement) among the row of \code{newdata}
(or the rows of the original data if \code{newdata} is not provided).

If \code{return.calib = TRUE}, \code{return.par = TRUE}
and/or \code{return.tau = TRUE}, then the list also contains respectively items
\code{calib}, \code{par} and/or \code{tau}.
}
\description{
Simulate from fitted \code{\link{gamBiCop-class}} object
}

