% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gamVine-class}
\alias{gamVine-class}
\title{The \code{\link{gamVine-class}}}
\description{
\code{\link{gamVine-class}} is an S4 class to store
 a generalized additive model on a vine copula.
}
\section{Slots}{

\describe{
\item{\code{Matrix}}{Lower triangular d x d matrix that defines the tree structure.}

\item{\code{model}}{list containing d x (d-1)/2 lists with three numeric items
(family, par and par2) and/or \code{\link{gamBiCop-class}} objects.}

\item{\code{names}}{vector of d names.}
}}
\seealso{
\code{\link{gamVine}}, \code{\link{RVineMatrix}} and
\code{\link{gamBiCop-class}}.
}

