% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gamVine}
\alias{gamVine}
\title{Constructor of the \code{\link{gamVine-class}}}
\usage{
gamVine(Matrix, model, names = NA)
}
\arguments{
\item{Matrix}{lower triangular d x d matrix that defines the tree structure.}

\item{model}{list containing d x (d-1)/2 lists with three numeric items
(family, par and par2) and/or \code{\link{gamBiCop-class}} objects.}

\item{names}{vector of d names.}
}
\value{
A \code{\link{gamVine-class}} object.
}
\description{
A constructor for objects of the \code{\link{gamVine-class}}.
}
\seealso{
\code{\link{gamVine-class}}, \code{\link{RVineMatrix}} and
\code{\link{gamBiCop-class}}.
}

