% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{gamVineCop-class}
\alias{gamVineCop-class}
\title{The \code{\link{gamVineCop-class}}}
\description{
\code{\link{gamVineCop-class}} is an S4 class to store
 a generalized additive model on a vine copula.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{lower triangular d x d matrix that defines the R-vine tree structure.}

\item{\code{model}}{list containing d x (d-1)/2 \code{\link{gamBiCop-class}} objects.}

\item{\code{names}}{vector of d names.}
}}
\seealso{
\code{\link{gamVineCop}}, \code{\link{RVineMatrix}} and \code{\link{gamBiCop-class}}.
}

