% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamVineCopFamily}
\alias{gamVineCopFamily}
\title{Family matrix of \code{\link{gamVineCop-class}} object}
\usage{
gamVineCopFamily(GVC)
}
\arguments{
\item{GVC}{\code{\link{gamVineCop-class}} object.}
}
\value{
Matrix of copula family (see \code{\link{gamBiCop-class}}) corresponding
to the model list in the \code{\link{gamVineCop-class}} object.
}
\description{
Return the matrix of copula family (see \code{\link{gamBiCop-class}}) corresponding
to the model list in the \code{\link{gamVineCop-class}} object.
}
\seealso{
\code{\link{gamVineCop-class}} and \code{\link{gamVineCop}}.
}

