% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamVineCopNormalize}
\alias{gamVineCopNormalize}
\title{Normalize a \code{\link{gamVineCop-class}} object}
\usage{
gamVineCopNormalize(GVC)
}
\arguments{
\item{GVC}{fitted \code{\link{gamVineCop-class}} object.}
}
\value{
Normalized \code{\link{gamVineCop-class}} object.
}
\description{
Change the R-vine matrix in the natural order,
i.e. with d:1 on the diagonal
}
\seealso{
\code{\link{gamVineCop-class}} and \code{\link{gamVineCop}}.
}

