% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamVineCopSim}
\alias{gamVineCopSim}
\title{Simulation from a \code{\link{gamVineCop-class}} object}
\usage{
gamVineCopSim(N, GVC, U = NULL)
}
\arguments{
\item{N}{number of d-dimensional observations to simulate.}

\item{GVC}{fitted \code{\link{gamVineCop-class}} object.}

\item{U}{(similar as \code{\link{RVineSim}} from the from the \code{\link{VineCopula}} package)
If not NULL, an (N,d)-matrix of U[0,1] random variates to be transformed to the copula sample.}
}
\value{
A Nxd matrix of data simulated from the given \code{\link{gamVineCop-class}} object.
}
\description{
Simulation from a \code{\link{gamVineCop-class}} object
}

