% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVine.R
\name{gamVineFamily}
\alias{gamVineFamily}
\title{Family Matrix of a gamVine Object}
\usage{
gamVineFamily(GVC)
}
\arguments{
\item{GVC}{A \code{\link[gamCopula:gamVine-class]{gamVine}} object.}
}
\value{
Matrix of copula families corresponding to the model list in the
\code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Return the matrix of copula family (see \code{\link{gamBiCop-class}})
corresponding to the model list in the \code{\link{gamVine-class}} object.
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}

