% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineMethods.R
\name{gamVineNormalize}
\alias{gamVineNormalize}
\title{Normalize a \code{\link{gamVine-class}} object}
\usage{
gamVineNormalize(GVC)
}
\arguments{
\item{GVC}{fitted \code{\link{gamVine-class}} object.}
}
\value{
Normalized \code{\link{gamVine-class}} object.
}
\description{
Change the R-vine matrix in the natural order,
i.e. with d:1 on the diagonal
}
\seealso{
\code{\link{gamVine-class}} and \code{\link{gamVine}}.
}

