% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineSeqEst.R
\name{gamVineSeqEst}
\alias{gamVineSeqEst}
\title{Sequential maximum penalized likelihood estimation of a GAM-Vine model.}
\usage{
gamVineSeqEst(data, GVC, covariates = NA, method = "FS", tol.rel = 0.001,
  n.iters = 10, verbose = FALSE)
}
\arguments{
\item{data}{A matrix or data frame containing the data in [0,1]^d.}

\item{GVC}{\code{\link{gamVine-class}} object.}

\item{covariates}{vector of names for the covariates.}

\item{method}{\code{'NR'} for Newton-Raphson
and  \code{'FS'} for Fisher-scoring (default).}

\item{tol.rel}{Relative tolerance for \code{'FS'}/\code{'NR'} algorithm.}

\item{n.iters}{Maximal number of iterations for
\code{'FS'}/\code{'NR'} algorithm.}

\item{verbose}{\code{TRUE} if informations should be printed during the
estimation and \code{FALSE} (default) for a silent version.}

\item{...}{Additional parameters to be passed to \code{\link{gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}.}
}
\value{
\code{gamVineSeqEst} returns a \code{\link{gamVine-class}} object.
}
\description{
This function estimates the parameter(s) of a Generalized Additive model
(GAM) Vine model, where GAMs for individual edges are specified either for
the copula parameter or Kendall's tau.
It solves the maximum penalized likelihood estimation for the copula families
supported in this package by reformulating each Newton-Raphson iteration as
a generalized ridge regression, which is solved using
the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\examples{
print(12)
}
\seealso{
\code{\link{gamVineCopSelect}} and
\code{\link{gamVineStructureSelect}}

\code{\link{gamVineCopSelect}},\code{\link{gamVineStructureSelect}},
 \code{\link{gamVine-class}}, \code{\link{gamVineSim}} and
 \code{\link{gamBiCopEst}}.
}

