% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gamVineSim}
\alias{gamVineSim}
\title{Simulation from a \code{\link{gamVine-class}} object}
\usage{
gamVineSim(N, GVC, U = NULL)
}
\arguments{
\item{N}{number of d-dimensional observations to simulate.}

\item{GVC}{\code{\link{gamVine-class}} object.}

\item{U}{(similar as \code{\link{RVineSim}} from the from the
\code{\link{VineCopula}} package) If not NULL, an (N,d)-matrix of U[0,1]
random variates to be transformed to the copula sample.}
}
\value{
A Nxd matrix of data simulated from the given
\code{\link{gamVine-class}} object.
}
\description{
Simulation from a \code{\link{gamVine-class}} object
}

