% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{logLik.GAMBiCop}
\alias{logLik,GAMBiCop-method}
\alias{logLik.GAMBiCop}
\title{Log-likelihood for a fitted \code{GAMBiCop} object}
\usage{
logLik.GAMBiCop(object, ...)

\S4method{logLik}{GAMBiCop}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
Standard \code{logLik} object: see \code{\link{logLik}}.
}
\description{
Function to extract the log-likelihood for a fitted \code{GAMBiCop-class}
object (note that the models are usually fitted by penalized likelihood maximization).
Used by \code{\link{AIC}} and \code{\link{BIC}}.
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

