% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{logLik.gamBiCop}
\alias{logLik,gamBiCop-method}
\alias{logLik.gamBiCop}
\title{Extract the Log-likelihood from a gamBiCop Object}
\usage{
\S4method{logLik}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{un-used in this class}
}
\value{
Standard \code{logLik} object: see \code{\link{logLik}}.
}
\description{
Function to extract the log-likelihood from an object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}} (note that the models are 
usually fitted by penalized likelihood maximization). 
This function is used by \code{\link{AIC}} and \code{\link{BIC}}.
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

