% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{nobs.GAMBiCop}
\alias{nobs,GAMBiCop-method}
\alias{nobs.GAMBiCop}
\title{Extract the Number of Obserations from a fitted \code{\link{GAMBiCop-class}} object}
\usage{
nobs.GAMBiCop(object, ...)

\S4method{nobs}{GAMBiCop}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{GAMBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
A single number, normally an integer.
}
\description{
Extract the number of 'observations' from a model fit.
This is principally intended to be used in computing BIC (see \code{\link{AIC}}).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

