% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{nobs.gamBiCop}
\alias{nobs,gamBiCop-method}
\alias{nobs.gamBiCop}
\title{Extract the Number of Obserations from gamBiCop Object}
\usage{
\S4method{nobs}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{un-used in this class}
}
\value{
A single number, normally an integer.
}
\description{
Extract the number of 'observations' from a model fit. 
This is principally intended to be used in computing the BIC 
(see \code{\link{AIC}}).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

