% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamBiCopMethods.R, R/gamVineMethods.R
\docType{methods}
\name{plot.gamBiCop}
\alias{plot.gamBiCop}
\alias{plot.gamVine}
\title{Plot a fitted \code{\link{gamBiCop-class}} object}
\usage{
plot.gamBiCop(x, ...)

plot.gamVine(x, ...)
}
\arguments{
\item{x}{fitted \code{\link{gamBiCop-class}} object.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}

\item{x}{fitted \code{\link{gamVine-class}} object.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}
}
\value{
This function simply generates plots.

This function simply generates plots.
}
\description{
Plot from a model fit.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).

Plot from a model fit.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\seealso{
\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).

\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).
}

