% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{plot.gamBiCop}
\alias{plot.gamBiCop}
\alias{plot,gamBiCop,ANY-method}
\title{Plot a gamBiCop Object}
\usage{
\S4method{plot}{gamBiCop,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{y}{Not used with this class.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}
}
\value{
This function simply generates plots.
}
\description{
Plot from an object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\seealso{
\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
