% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVine.R
\docType{methods}
\name{plot.gamVine}
\alias{plot,gamVine,ANY-methods}
\alias{plot.gamVine}
\title{Plot a gamVine Class Object}
\usage{
\S4method{plot}{gamVine,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.}

\item{y}{Not used with this class.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}
}
\value{
This function simply generates plots.
}
\description{
Plot from an object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\seealso{
\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).
}

