% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print.gamVineCop}
\alias{print,gamVineCop-method}
\alias{print.gamVineCop}
\title{Print a \code{gamVineCop-class} object}
\usage{
print.gamVineCop(x, detail = FALSE, ...)

\S4method{print}{gamVineCop}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{fitted \code{\link{gamVineCop-class}} object.}

\item{detail}{should additional details be printed (\code{detail=TRUE}) or not?}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a \code{gamVineCop-class} object.
}
\seealso{
\code{gamVineCop-class} and \code{gamVineCop}.
}

