% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RVM2GVC}
\alias{RVM2GVC}
\title{\code{\link{RVineMatrix}} to \code{\link{gamVineCop-class}}}
\usage{
RVM2GVC(RVM)
}
\arguments{
\item{RVM}{\code{\link{RVineMatrix}} object.}
}
\value{
A \code{\link{gamVineCop-class}} object.
}
\description{
Transform a \code{\link{RVineMatrix}} object
into a \code{\link{gamVineCop-class}} object.
}
\seealso{
\code{\link{RVineMatrix}}, \code{\link{gamVineCop-class}} and \code{\link{gamVineCop}}.
}

