% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{summary.gamBiCop}
\alias{summary.gamBiCop}
\alias{summary,gamBiCop-method}
\title{Summary for a gamBiCop Object}
\usage{
\S4method{summary}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{unused in this class}
}
\value{
A useful summary (see \code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}} for more details).
}
\description{
Takes a \code{\link[gamCopula:gamBiCop-class]{gamBiCop}} object and produces
various useful summaries from it.
}
\seealso{
\code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}
}
