% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVine.R
\docType{methods}
\name{summary.gamVine}
\alias{summary,gamVine-method}
\alias{summary.gamVine}
\title{Summary for a gamVine Class Oject}
\usage{
\S4method{summary}{gamVine}(object)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.}

\item{...}{un-used in this class}
}
\value{
A useful summary (see \code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}} for more details).
}
\description{
Takes a \code{\link{gamVine-class}} object and produces various
useful summaries from it.
}
\seealso{
\code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}
}

