% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{BIC.gamBiCop}
\alias{BIC,gamBiCop-method}
\alias{BIC.gamBiCop}
\title{Schwarz's Bayesian Information Criterion for a fitted \code{\link{gamBiCop-class}}}
\usage{
BIC.gamBiCop(object, ...)

\S4method{BIC}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{gamBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
A numeric value with the corresponding BIC.
}
\description{
Function calculating the Schwarz's Bayesian Information Criterion (BIC)
for a fitted \code{\link{gamBiCop-class}} object
(note that the models are usually fitted by penalized likelihood maximization).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

