% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineMethods.R
\docType{methods}
\name{dim.gamVine}
\alias{dim,gamVine-method}
\alias{dim.gamVine}
\title{Dimension of a \code{\link{gamVine-class}} object}
\usage{
\method{dim}{gamVine}(x)

\S4method{dim}{gamVine}(x)
}
\arguments{
\item{x}{fitted \code{\link{gamVine-class}} object.}
}
\value{
Dimension of the \code{\link{gamVine-class}} object.
}
\description{
Retrieve the dimension of a \code{\link{gamVine-class}} object.
}
\seealso{
\code{\link{gamVine-class}} and \code{\link{gamVine}}.
}

