% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{EDF}
\alias{EDF}
\title{Equivalent Degrees of Freedom for an Object of the Class gamBiCop}
\usage{
EDF(object)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}
}
\value{
Estimated degrees of freedom for each smooth component.
}
\description{
Function calculating the Equivalent Degrees of Freedom (EDF)
for a \code{\link[gamCopula:gamBiCop-class]{gamBiCop}} object.
It basically sums the edf of the \code{\link[mgcv]{gamObject}}
for each smooth component.
}
