% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{formula.gamBiCop}
\alias{formula,gamBiCop-method}
\alias{formula.gamBiCop}
\title{\code{\link{gamBiCop-class}} formula}
\usage{
formula.gamBiCop(x, ...)

\S4method{formula}{gamBiCop}(x, ...)
}
\arguments{
\item{x}{fitted \code{\link{gamBiCop-class}} object.}

\item{...}{un-used in this class}
}
\description{
Description of the \code{\link{gam}} formula for  fitted \code{\link{gamBiCop-class}} object.
This function is a wrapper to \code{\link{formula.gam}}
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\seealso{
\code{\link{formula.gam}} function
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}

