% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamBiCop.R
\docType{methods}
\name{formula.gamBiCop}
\alias{formula.gamBiCop}
\alias{formula,gamBiCop-method}
\title{Model Formula of the gamBiCop Object}
\usage{
\S4method{formula}{gamBiCop}(x, ...)
}
\arguments{
\item{x}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{un-used in this class}
}
\description{
Extracts the \code{\link[mgcv]{gam}} formula from an object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.
This function is a wrapper to \code{\link[mgcv]{formula.gam}}
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\seealso{
\code{\link[mgcv]{formula.gam}} function
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
