% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gamBiCop-class}
\alias{gamBiCop-class}
\title{The gamBiCop Class}
\description{
gamBiCop is an S4 class to store
 a Generalized Additive Model for bivariate copula a parameter or
 Kendall's tau. Objects can be created by calls of the form
 \code{new("gamBiCop", ...)}, or by function \code{\link{gamBiCop}}.
}
\section{Slots}{

\describe{
\item{\code{family}}{A copula family: \code{1} Gaussian,
\code{2} Student t,
\code{3} Clayton,
\code{4} Gumbel,
\code{13} Survival Clayton,
\code{14} Survival Gumbel,
\code{23} Rotated (90 degrees) Clayton,
\code{24} Rotated (90 degrees) Gumbel,
\code{33} Rotated (270 degrees) Clayton and
\code{34} Rotated (270 degrees) Gumbel.}

\item{\code{model}}{A \code{\link{gamObject}} as return by the
 \code{\link{gam}} function
from the \code{\link[mgcv:mgcv-package]{mgcv}} package.}

\item{\code{par2}}{Second parameter for the Studen t-copula.}

\item{\code{tau}}{\code{FALSE} (default) for a calibration fonction
specified for the Copula parameter
or \code{TRUE} for a calibration function specified for Kendall's tau.}
}}
\seealso{
\code{\link{gamBiCopEst}},
\code{\link{gamBiCopPred}} and \code{\link{gamBiCopSim}}.
}

