% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\name{gamVine}
\alias{gamVine}
\title{Construction of a gamVine Class Object}
\usage{
gamVine(Matrix, model, names = NA, covariates = NA)
}
\arguments{
\item{Matrix}{lower triangular d x d matrix that defines the tree structure.}

\item{model}{list containing d x (d-1)/2 lists with three numeric items
(family, par and par2) and/or objects of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{names}{vector of d names.}

\item{covariates}{vector of names for the covariates.}
}
\value{
An object of the class
 \code{\link[gamCopula:gamVine-class]{gamVine}}.
}
\description{
Constructs an object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}},
\code{\link[VineCopula]{RVineMatrix}}, \code{\link[gamCopula:gamBiCop-class]{gamBiCop}}
\code{\link{gamVineSeqFit}}, \code{\link{gamVineCopSelect}},
\code{\link{gamVineStructureSelect}} and \code{\link{gamVineSimulate}}.
}
