% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamVineCopFamily}
\alias{gamVineCopFamily}
\title{Family matrix of \code{gamVineCop-class} object}
\usage{
gamVineCopFamily(x)
}
\arguments{
\item{x}{fitted \code{\link{gamVineCop-class}} object.}
}
\value{
Matrix of copula family (see \code{\link{gamBiCop-class}}) corresponding
to the model list in the \code{gamVineCop-class} object.
}
\description{
Return the matrix of copula family (see \code{\link{gamBiCop-class}}) corresponding
to the model list in the \code{gamVineCop-class} object.
}
\seealso{
\code{gamVineCop-class} and \code{gamVineCop}.
}

