% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineMethods.R
\name{gamVineFamily}
\alias{gamVineFamily}
\title{Family matrix of \code{\link{gamVine-class}} object}
\usage{
gamVineFamily(GVC)
}
\arguments{
\item{GVC}{\code{\link{gamVine-class}} object.}
}
\value{
Matrix of copula family (see \code{\link{gamBiCop-class}})
corresponding to the model list in the \code{\link{gamVine-class}} object.
}
\description{
Return the matrix of copula family (see \code{\link{gamBiCop-class}})
corresponding to the model list in the \code{\link{gamVine-class}} object.
}
\seealso{
\code{\link{gamVine-class}} and \code{\link{gamVine}}.
}

