% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\name{gamVineFamily}
\alias{gamVineFamily}
\title{Family Matrix of an Object of the Class gamVine}
\usage{
gamVineFamily(GVC)
}
\arguments{
\item{GVC}{An object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.}
}
\value{
Matrix of copula families corresponding to the model list in the 
\code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Return the matrix of copula family (see 
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}) 
corresponding to the model list in the 
\code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}
