% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVine.R
\name{gamVineNormalize}
\alias{gamVineNormalize}
\title{Normalize a gamVine Object}
\usage{
gamVineNormalize(GVC)
}
\arguments{
\item{GVC}{A \code{\link[gamCopula:gamVine-class]{gamVine}} object.}
}
\value{
The normalized \code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Change the R-vine matrix in the natural order,
i.e. with d:1 on the diagonal
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}

