% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\name{gamVineNormalize}
\alias{gamVineNormalize}
\title{Normalize an Object of the Class gamVine}
\usage{
gamVineNormalize(GVC)
}
\arguments{
\item{GVC}{An object of the class 
\code{\link[gamCopula:gamVine-class]{gamVine}}.}
}
\value{
The normalized \code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Change the R-vine matrix in the natural order, 
i.e. with d:1 on the diagonal
}
\seealso{
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}

