% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineSeqEst.R
\name{gamVineSeqEst}
\alias{gamVineSeqEst}
\title{Sequential maximum penalized likelihood estimation of a GAM-Vine model.}
\usage{
gamVineSeqEst(data, GVC, method = "NR", tol.rel = 0.001, n.iters = 10,
  verbose = FALSE)
}
\arguments{
\item{data}{A matrix or data frame containing the data in [0,1]^d.}

\item{GVC}{\code{\link{gamVine-class}} object.}

\item{method}{\code{'NR'} for Newton-Raphson
and  \code{'FS'} for Fisher-scoring (default).}

\item{tol.rel}{Relative tolerance for \code{'FS'}/\code{'NR'} algorithm.}

\item{n.iters}{Maximal number of iterations for
\code{'FS'}/\code{'NR'} algorithm.}

\item{verbose}{\code{TRUE} if informations should be printed during the
estimation and \code{FALSE} (default) for a silent version.}

\item{...}{Additional parameters to be passed to \code{\link{gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}.}
}
\value{
\code{gamVineSeqEst} returns a \code{\link{gamVine-class}} object.
}
\description{
This function estimates the parameter(s) of a Generalized Additive model
(GAM) Vine model, where GAMs for individual edges are specified either for
the copula parameter or Kendall's tau.
It solves the maximum penalized likelihood estimation for the copula families
supported in this package by reformulating each Newton-Raphson iteration as
a generalized ridge regression, which is solved using
the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\examples{
require(VineCopula)
set.seed(0)

## A first example with a 3-dimensional GAM-Vine

# Define a R-vine tree structure matrix
d <- 3
Matrix <- c(2,3,1,0,3,1,0,0,1)
Matrix <- matrix(Matrix,d,d)
nnames <- paste("x", 1:d, sep = "")

# Copula families for each edge
fam <- c(3,4,1)

# Parameters for the first tree (two unconditional copulas)
par <- c(1,2)

# Pre-allocate the GAM-Vine model list
count <- 1
model <- vector(mode = "list", length = d*(d-1)/2)

# The first tree contains only the two unconditional copulas
for (i in 1:(d-1)) {
  model[[count]] <- list(family = fam[count], par = par[count], par2 = 0)
  count <- count + 1
}

# The second tree contains a unique conditional copula
# In this first example, we take a linear calibration function (10*x-5)

# Set-up a dummy data
tmp <- data.frame(u1 = runif(1e2), u2 = runif(1e2), x1 = runif(1e2))

# Set-up an arbitrary linear model for the calibration function
model[[count]] <- gamBiCopEst(tmp, ~ x1, fam[count])$res

# Update the coefficients of the model
attr(model[[count]],"model")$coefficients <- c(-5, 10)

# Define gamVine object
GVC <- gamVine(Matrix = Matrix, model = model, names = nnames)
GVC

# Simulate new data
N <- 1e3
simData <- data.frame(gamVineSim(N, GVC))
colnames(simData) <- nnames

# Fit data
summary(fitGVC <- gamVineSeqEst(simData, GVC))

# The second tree contains a unique conditional copula
# In this second example, we take a smooth calibration function

# Set-up an arbitrary smooth model with 10 knots for the calibration function
model[[count]] <- gamBiCopEst(tmp, ~ s(x1, k = 10, fx = T),
                              fam[count], n.iters = 1)$res

# Update the coefficients of the model
attr(model[[count]],"model")$coefficients <- rnorm(11)
plot(attr(model[[count]],"model"), se = F)

# Update the gamVine object
GVC <- gamVine(Matrix = Matrix, model = model, names = nnames)
summary(GVC)

# Simulate new data
N <- 1e3
simData <- data.frame(gamVineSim(N, GVC))
colnames(simData) <- nnames

# Fit data
summary(fitGVC <- gamVineSeqEst(simData, GVC))
plot(fitGVC)
}
\seealso{
\code{\link{gamVine-class}}, \code{\link{gamVineSim}} and
\code{\link{gamBiCopEst}}.
}

