% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineStructureSelect.R
\name{gamVineStructureSelect}
\alias{gamVineStructureSelect}
\title{Structure selection and estimation of a GAM-Vine model.}
\usage{
gamVineStructureSelect(data, type = 0, familyset = NA, rotations = TRUE,
  selectioncrit = "AIC", SAtestOptions = "ERC", indeptest = TRUE,
  level = 0.05, trunclevel = NA, tau = TRUE, method = "FS",
  tol.rel = 0.001, n.iters = 10, verbose = FALSE)
}
\arguments{
\item{data}{A matrix or data frame containing the data in [0,1]^d.}

\item{type}{\code{type = 0} (default) for a R-Vine and \code{type = 1} for a
C-Vine}

\item{familyset}{An integer vector of pair-copula families to select from
(the independence copula MUST NOT be specified in this vector unless one
wants to fit an independence vine!). The vector has to include at least one
pair-copula family that allows for positive and one that allows for negative
dependence. Not listed copula families might be included to better handle
limit cases. If \code{familyset = NA} (default), selection among all
possible families is performed. Coding of pair-copula families:
\code{1} Gaussian, \code{2} Student t,
\code{3} Clayton, \code{4} Gumbel, \code{13} Survival Clayton,
\code{14} Survival Gumbel,  \code{23} Rotated (90 degrees) Clayton,
\code{24} Rotated (90 degrees) Gumbel,
\code{33} Rotated (270 degrees) Clayton and
\code{34} Rotated (270 degrees) Gumbel.}

\item{rotations}{If \code{TRUE}, all rotations of the families in familyset
are included.}

\item{selectioncrit}{Character indicating the criterion for bivariate copula
selection. Possible choices: \code{selectioncrit = 'AIC'} (default) or
\code{'BIC'}, as in \code{\link{BiCopSelect}} from the
\code{\link[VineCopula:VineCopula-package]{VineCopula}} package.}

\item{SAtestOptions}{TODO;TODO;TODO;TODO;TODO;TODO;TODO!!!}

\item{indeptest}{Logical; whether a hypothesis test for the simplifying
assumption and the independence of
\code{u1} and \code{u2} is performed before bivariate copula selection
(default: \code{indeptest = TRUE}; see \code{\link{BiCopIndTest}} and
\code{\link{SAtest}}).
The independence copula is chosen for a (conditional) pair if both the
simplifying assumption and the null
hypothesis of independence cannot be rejected.}

\item{level}{Numerical; significance level of the test (default:
\code{level = 0.05}).}

\item{trunclevel}{Integer; level of truncation.}

\item{tau}{\code{TRUE} (default) for a calibration fonction specified for
Kendall's tau or \code{FALSE} for a calibration function specified
for the Copula parameter.}

\item{method}{\code{'NR'} for Newton-Raphson
and  \code{'FS'} for Fisher-scoring (default).}

\item{tol.rel}{Relative tolerance for \code{'FS'}/\code{'NR'} algorithm.}

\item{n.iters}{Maximal number of iterations for
\code{'FS'}/\code{'NR'} algorithm.}

\item{verbose}{\code{TRUE} if informations should be printed during the
estimation and \code{FALSE} (default) for a silent version.
from \code{\link[mgcv:mgcv-package]{mgcv}}.}
}
\value{
\code{gamVineSeqEst} returns a \code{\link{gamVine-class}} object.
}
\description{
This function select the structure and estimates the parameter(s) of a
Generalized Additive model
(GAM) Vine model, where GAMs for individual edges are specified either for
the copula parameter or Kendall's tau.
It solves the maximum penalized likelihood estimation for the copula families
supported in this package by reformulating each Newton-Raphson iteration as
a generalized ridge regression, which is solved using
the \code{\link[mgcv:mgcv-package]{mgcv}} package.
}
\examples{
require(VineCopula)
set.seed(0)

## A first example with a 3-dimensional GAM-Vine

# Define a R-vine tree structure matrix
d <- 3
Matrix <- c(2,3,1,0,3,1,0,0,1)
Matrix <- matrix(Matrix,d,d)
nnames <- paste("x", 1:d, sep = "")

# Copula families for each edge
fam <- c(3,4,1)

# Parameters for the first tree (two unconditional copulas)
par <- c(1,2)

# Pre-allocate the GAM-Vine model list
count <- 1
model <- vector(mode = "list", length = d*(d-1)/2)

# The first tree contains only the two unconditional copulas
for (i in 1:(d-1)) {
  model[[count]] <- list(family = fam[count], par = par[count], par2 = 0)
  count <- count + 1
}

# The second tree contains a unique conditional copula
# In this first example, we take a linear calibration function (10*x-5)

# Set-up a dummy dataset
tmp <- data.frame(u1 = runif(1e2), u2 = runif(1e2), x1 = runif(1e2))

# Set-up an arbitrary linear model for the calibration function
model[[count]] <- gamBiCopEst(tmp, ~ x1, fam[count])$res

# Update the coefficients of the model
attr(model[[count]],"model")$coefficients <- c(-5, 10)

# Define gamVine object
GVC <- gamVine(Matrix = Matrix, model = model, names = nnames)
GVC

# Simulate new data
N <- 1e3
simData <- data.frame(gamVineSim(N, GVC))
colnames(simData) <- nnames

# Fit data
summary(fitGVC <- gamVineSeqEst(simData, GVC))
summary(fitGVC2 <- gamVineStructureSelect(simData, GVC))
}
\seealso{
\code{\link{gamVine-class}}, \code{\link{gamVineSim}},
\code{\link{gamVineSeqEst}} and \code{\link{gamBiCopEst}}.
}

