% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{logLik.gamBiCop}
\alias{logLik,gamBiCop-method}
\alias{logLik.gamBiCop}
\title{Log-likelihood for a \code{\link{gamBiCop-class}} object}
\usage{
logLik.gamBiCop(object, ...)

\S4method{logLik}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{\code{\link{gamBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
Standard \code{logLik} object: see \code{\link{logLik}}.
}
\description{
Function to extract the log-likelihood for a fitted \code{\link{gamBiCop-class}}
object (note that the models are usually fitted by penalized likelihood maximization).
This function is used by \code{\link{AIC}} and \code{\link{BIC}}.
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

