% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{nobs.gamBiCop}
\alias{nobs,gamBiCop-method}
\alias{nobs.gamBiCop}
\title{Extract the number of obserations from a fitted \code{\link{gamBiCop-class}} object}
\usage{
nobs.gamBiCop(object, ...)

\S4method{nobs}{gamBiCop}(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{gamBiCop-class}} object.}

\item{...}{un-used in this class}
}
\value{
A single number, normally an integer.
}
\description{
Extract the number of 'observations' from a model fit.
This is principally intended to be used in computing BIC (see \code{\link{AIC}}).
}
\seealso{
\code{\link{AIC}} and \code{\link{BIC}}.
}

