% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamBiCop.R, R/gamVine.R
\docType{methods}
\name{plot.gamBiCop}
\alias{plot,gamVine,ANY-method}
\alias{plot.gamBiCop}
\title{Plot a gamBiCop Object}
\usage{
\S4method{plot}{gamBiCop,ANY}(x, y, ...)

\S4method{plot}{gamVine,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}

\item{x}{An object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}}

\item{...}{additional arguments to be passed to \code{\link{plot.gam}}.}
}
\value{
This function simply generates plots.

This function simply generates plots.
}
\description{
Plot from an object of the class
\code{\link[gamCopula:gamBiCop-class]{gamBiCop}}.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).

Plot from an object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\seealso{
\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).

\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).
}

