% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.gamBiCop}
\alias{plot.gamBiCop}
\title{Plot a fitted \code{\link{gamBiCop-class}} object}
\usage{
plot.gamBiCop(x, ...)
}
\arguments{
\item{x}{fitted \code{\link{gamBiCop-class}} object.}

\item{...}{additional arguments to be passed to plot.gam.}
}
\value{
This function simply generates plots.
}
\description{
Plot from a model fit.
The function is based on (see \code{\link{plot.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}).
}
\seealso{
\code{\link{plot.gam}} from \code{\link[mgcv:mgcv-package]{mgcv}}).
}

