% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print.gamVine}
\alias{print,gamVine-method}
\alias{print.gamVine}
\title{Print a \code{\link{gamVine-class}} object}
\usage{
print.gamVine(x, ...)

\S4method{print}{gamVine}(x, ...)
}
\arguments{
\item{x}{\code{\link{gamVine-class}} object.}

\item{...}{un-used for this class.}
}
\description{
Print a \code{\link{gamVine-class}} object
}
\seealso{
\code{\link{gamVine-class}} and \code{\link{gamVine}}.
}

