% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVine.R
\name{RVM2GVC}
\alias{RVM2GVC}
\title{Transform an R-Vine Class Object into a gamVine Class Object}
\usage{
RVM2GVC(RVM)
}
\arguments{
\item{RVM}{A \code{\link{RVineMatrix}} object.}
}
\value{
A \code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\description{
Transform a \code{\link{RVineMatrix}} object
into a \code{\link[gamCopula:gamVine-class]{gamVine}} object.
}
\seealso{
\code{\link{RVineMatrix}} and
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}

