% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamVineMethods.R
\name{RVM2GVC}
\alias{RVM2GVC}
\title{\code{\link{RVineMatrix}} to \code{\link{gamVine-class}}}
\usage{
RVM2GVC(RVM)
}
\arguments{
\item{RVM}{\code{\link{RVineMatrix}} object.}
}
\value{
A \code{\link{gamVine-class}} object.
}
\description{
Transform a \code{\link{RVineMatrix}} object
into a \code{\link{gamVine-class}} object.
}
\seealso{
\code{\link{RVineMatrix}}, \code{\link{gamVine-class}} and
\code{\link{gamVine}}.
}

