% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\name{RVM2GVC}
\alias{RVM2GVC}
\title{Transform an Object of the Class R-Vine into an Object of the Class gamVine}
\usage{
RVM2GVC(RVM)
}
\arguments{
\item{RVM}{An object of the class \code{\link{RVineMatrix}}.}
}
\value{
An object of the class 
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}
\description{
Transform an object of the class \code{\link{RVineMatrix}}
into an object of the class \code{\link[gamCopula:gamVine-class]{gamVine}}.
}
\seealso{
\code{\link{RVineMatrix}} and 
\code{\link[gamCopula:gamVine-class]{gamVine}}.
}
