% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamVine.R
\docType{methods}
\name{summary.gamVine}
\alias{summary.gamVine}
\alias{summary,gamVine-method}
\title{Summary for an Object of the Class gamVine}
\usage{
\S4method{summary}{gamVine}(object)
}
\arguments{
\item{object}{An object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}}.}

\item{...}{unused in this class}
}
\value{
A useful summary (see \code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}} for more details).
}
\description{
Takes an object of the class
\code{\link[gamCopula:gamVine-class]{gamVine}} and produces various
useful summaries from it.
}
\seealso{
\code{\link{summary.gam}}
from \code{\link[mgcv:mgcv-package]{mgcv}}
}
